requirements-parakeet.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchaudio
nemo_toolkit[asr]
